requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
bcrypt
pyjwt
python-dotenv
uv
psycopg2-binary